Add GetFirstValueOnTableAsInteger and GetFirstValueOnDataSet. Move data set extensions from data  table t data set extensions class. Fix the GetNumberOfRowsOnTable method. It was returning the first value on the table rows/column.
